+++ src/app/pages/elements/header/header.component.html
@@ -1,5 +1,5 @@
 <doc-page
-    header="Headers"
+    header="Header"
     subHeader="A header provides a short summary of content">
   <div *docPageContent="'definition'">
     <h2 sui-header
