+++ src/app/pages/elements/icon/icon.component.scss
