+++ src/app/pages/elements/icon/icon.component.ts
